chore: refactor workflow run table
